removes more bunny gem and extracts 'sluggit' into it's own ruby file
